# feat(tiny_ttf): add kerning cache