--- a/src/lv_conf_internal.h
+++ b/src/lv_conf_internal.h
@@ -3061,7 +3061,14 @@
         #ifdef CONFIG_LV_TINY_TTF_CACHE_GLYPH_CNT
             #define LV_TINY_TTF_CACHE_GLYPH_CNT CONFIG_LV_TINY_TTF_CACHE_GLYPH_CNT
         #else
-            #define LV_TINY_TTF_CACHE_GLYPH_CNT 256
+            #define LV_TINY_TTF_CACHE_GLYPH_CNT 128
+        #endif
+    #endif
+    #ifndef LV_TINY_TTF_CACHE_KERNING_CNT
+        #ifdef CONFIG_LV_TINY_TTF_CACHE_KERNING_CNT
+            #define LV_TINY_TTF_CACHE_KERNING_CNT CONFIG_LV_TINY_TTF_CACHE_KERNING_CNT
+        #else
+            #define LV_TINY_TTF_CACHE_KERNING_CNT 256
         #endif
     #endif
 #endif