Performance measured using this example code:

```c
	static lv_style_t style;
	lv_style_init(&style);
	lv_font_t *font = lv_tiny_ttf_create_file("A:myriadpro.ttf", 30);
	lv_style_set_text_font(&style, font);
	lv_style_set_text_align(&style, LV_TEXT_ALIGN_CENTER);

	/*Create a label with the new style*/
	lv_obj_t *label = lv_label_create(lv_screen_active());
	lv_obj_add_style(label, &style, 0);
	for (size_t count = 1; count < 20; ++count) {
		lv_label_set_text(
			label,
			"Lorem ipsum dolor sit amet consectetur adipiscing elit."
			"Quisque faucibus ex sapien vitae pellentesque sem placerat."
			"In id cursus mi pretium tellus duis convallis."
			"Tempus leo eu aenean sed diam urna tempor."
			"Pulvinar vivamus fringilla lacus nec metus bibendum egestas."
			"Iaculis massa nisl malesuada lacinia integer nunc posuere."
			"Ut hendrerit semper vel class aptent taciti sociosqu."
			"Ad litora torquent per conubia nostra inceptos himenaeos."
			"Lorem ipsum dolor sit amet consectetur adipiscing elit."
			"Quisque faucibus ex sapien vitae pellentesque sem placerat."
			"In id cursus mi pretium tellus duis convallis."
			"Tempus leo eu aenean sed diam urna tempor."
			"Pulvinar vivamus fringilla lacus nec metus bibendum egestas."
			"Iaculis massa nisl malesuada lacinia integer nunc posuere."
			"Ut hendrerit semper vel class aptent taciti sociosqu."
			"Ad litora torquent per conubia nostra inceptos himenaeos."
			"Lorem ipsum dolor sit amet consectetur adipiscing elit."
			"Quisque faucibus ex sapien vitae pellentesque sem placerat."
			"In id cursus mi pretium tellus duis convallis."
			"Tempus leo eu aenean sed diam urna tempor."
			"Pulvinar vivamus fringilla lacus nec metus bibendum egestas."
			"Iaculis massa nisl malesuada lacinia integer nunc posuere."
			"Ut hendrerit semper vel class aptent taciti sociosqu."
			"Ad litora torquent per conubia nostra inceptos himenaeos."
			"Lorem ipsum dolor sit amet consectetur adipiscing elit."
			"Quisque faucibus ex sapien vitae pellentesque sem placerat."
			"In id cursus mi pretium tellus duis convallis."
			"Tempus leo eu aenean sed diam urna tempor."
			"Pulvinar vivamus fringilla lacus nec metus bibendum egestas."
			"Iaculis massa nisl malesuada lacinia integer nunc posuere."
			"Ut hendrerit semper vel class aptent taciti sociosqu."
			"Ad litora torquent per conubia nostra inceptos himenaeos."
			"Lorem ipsum dolor sit amet consectetur adipiscing elit."
			"Quisque faucibus ex sapien vitae pellentesque sem placerat."
			"In id cursus mi pretium tellus duis convallis."
			"Tempus leo eu aenean sed diam urna tempor."
			"Pulvinar vivamus fringilla lacus nec metus bibendum egestas."
			"Iaculis massa nisl malesuada lacinia integer nunc posuere."
			"Ut hendrerit semper vel class aptent taciti sociosqu."
			"Ad litora torquent per conubia nostra inceptos himenaeos."
			"Lorem ipsum dolor sit amet consectetur adipiscing elit."
			"Quisque faucibus ex sapien vitae pellentesque sem placerat."
			"In id cursus mi pretium tellus duis convallis."
			"Tempus leo eu aenean sed diam urna tempor."
			"Pulvinar vivamus fringilla lacus nec metus bibendum egestas."
			"Iaculis massa nisl malesuada lacinia integer nunc posuere."
			"Ut hendrerit semper vel class aptent taciti sociosqu."
			"Ad litora torquent per conubia nostra inceptos himenaeos.");
	}

	lv_obj_center(label);
	lv_refr_now(NULL); /*Update layouts and render*/
	lv_obj_invalidate(lv_screen_active());

	const uint32_t t = lv_tick_get();
	lv_refr_now(NULL); /*Render only*/
	LV_LOG_USER("%u ms\n", lv_tick_elaps(t));

	return 0;
```

Measured with `LV_TINY_TTF_CACHE_KERNING_CNT == 0` (no caching) and `LV_TINY_TTF_CACHE_KERNING_CNT == 256` (so extra 3072 bytes in RAM)

Application start time reduced from 1.62s to 236ms
Render time improved from ~95ms to ~7ms
Flamegraph shows kerning value fetch dropped from 50% to just 0.91% of total CPU cycles

![Screenshot From 2025-05-27 16-07-51](https://github.com/user-attachments/assets/2e5e9c08-f32e-43d5-8778-61327e261b29)
![Screenshot From 2025-05-27 16-10-33](https://github.com/user-attachments/assets/d7ddfc9c-5726-45f6-8ce5-8c50314fd35b)
![Screenshot From 2025-05-27 16-14-48](https://github.com/user-attachments/assets/eb46ec6d-aa4e-4520-a426-03edf99fe405)