@AndreCostaaa 2025-05-28
> Please also update Kconfig.

Done !