@AndreCostaaa 2025-06-04
FYI - after investigating a bit, the actual cache size needed to store every glyph pair in this example is 183 so that's why after reaching 255 the performance stays the same